requirements.txt
cachelib==0.1.1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Click==7.0
Flask==1.0.2
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
packaging==20.9
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.23
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==0.14.1
